+++ projects/rdc/rdc_libs/rdc/src/SmiUtils.cc
@@ -77,6 +77,18 @@ rdc_status_t Smi2RdcError(amdsmi_status_t rsmi) {
   }
 }
 
+bool is_capability_miss(rdc_status_t status) {
+  switch (status) {
+    case RDC_ST_NOT_SUPPORTED:
+    case RDC_ST_BAD_PARAMETER:  // AMDSMI_STATUS_INVAL, e.g. xgmi_error read on gfx950
+    case RDC_ST_NOT_FOUND:
+    case RDC_ST_PERM_ERROR:  // privileges do not change for the life of the daemon
+      return true;
+    default:
+      return false;
+  }
+}
+
 amdsmi_status_t get_processor_handle_from_id(uint32_t gpu_id,
                                              amdsmi_processor_handle* processor_handle) {
   uint32_t socket_count = 0;